test(Summary): add tests for content stacking in mobile
